+++ test/Sentry.Serilog.Tests/SentrySerilogSinkExtensionsTests.cs
@@ -28,6 +28,7 @@ public class SentrySerilogSinkExtensionsTests
         public bool InitializeSdk { get; } = false;
         public LogEventLevel MinimumEventLevel { get; } = LogEventLevel.Verbose;
         public LogEventLevel MinimumBreadcrumbLevel { get; } = LogEventLevel.Fatal;
+        public bool ExperimentalEnableLogs { get; } = true;
 
         public static SentrySerilogOptions GetSut() => new();
     }
@@ -97,7 +98,7 @@ public class SentrySerilogSinkExtensionsTests
             _fixture.SampleRate, _fixture.Release, _fixture.Environment, _fixture.MaxQueueItems,
             _fixture.ShutdownTimeout, _fixture.DecompressionMethods, _fixture.RequestBodyCompressionLevel,
             _fixture.RequestBodyCompressionBuffered, _fixture.Debug, _fixture.DiagnosticLevel,
-            _fixture.ReportAssembliesMode, _fixture.DeduplicateMode);
+            _fixture.ReportAssembliesMode, _fixture.DeduplicateMode, null, _fixture.ExperimentalEnableLogs);
 
         // Compare individual properties
         Assert.Equal(_fixture.SendDefaultPii, sut.SendDefaultPii);
@@ -108,7 +109,7 @@ public class SentrySerilogSinkExtensionsTests
         Assert.Equal(_fixture.SampleRate, sut.SampleRate);
         Assert.Equal(_fixture.Release, sut.Release);
         Assert.Equal(_fixture.Environment, sut.Environment);
-        Assert.Equal(_fixture.Dsn, sut.Dsn!);
+        Assert.Equal(_fixture.Dsn, sut.Dsn);
         Assert.Equal(_fixture.MaxQueueItems, sut.MaxQueueItems);
         Assert.Equal(_fixture.ShutdownTimeout, sut.ShutdownTimeout);
         Assert.Equal(_fixture.DecompressionMethods, sut.DecompressionMethods);
@@ -118,6 +119,7 @@ public class SentrySerilogSinkExtensionsTests
         Assert.Equal(_fixture.DiagnosticLevel, sut.DiagnosticLevel);
         Assert.Equal(_fixture.ReportAssembliesMode, sut.ReportAssembliesMode);
         Assert.Equal(_fixture.DeduplicateMode, sut.DeduplicateMode);
+        Assert.Equal(_fixture.ExperimentalEnableLogs, sut.Experimental.EnableLogs);
         Assert.True(sut.InitializeSdk);
         Assert.Equal(_fixture.MinimumEventLevel, sut.MinimumEventLevel);
         Assert.Equal(_fixture.MinimumBreadcrumbLevel, sut.MinimumBreadcrumbLevel);